Implement SVD based training step & f64 floats

+++ src/activation.rs
@@ -12,7 +12,7 @@ pub enum Activation {
 }
 
 impl Activation {
-    pub fn apply(&self, input: f32) -> f32 {
+    pub fn apply(&self, input: f64) -> f64 {
         match self {
             Activation::ReLu => relu(input),
             Activation::Sigmoid => sigmoid(input),
@@ -21,14 +21,14 @@ impl Activation {
     }
 }
 
-pub fn relu(input: f32) -> f32 {
+pub fn relu(input: f64) -> f64 {
     input.max(0.0)
 }
 
-pub fn sigmoid(input: f32) -> f32 {
+pub fn sigmoid(input: f64) -> f64 {
     1.0 / (1.0 + (-input).exp())
 }
 
-pub fn tanh(input: f32) -> f32 {
+pub fn tanh(input: f64) -> f64 {
     2.0 / (1.0 + (-2.0 * input).exp()) - 1.0
 }
